+++ Source/math.h
@@ -1,3 +1,3 @@
 ﻿#pragma once
 
-void get_slope(int x1, int y1, int x2, int y2, int *dx, int *dy);
+void get_slope(int x1, int y1, int x2, int y2, float *dx, float *dy);
